# zwf-app/src/_assets/api.yml
swagger: "2.0"
info:
  description: ZWF backend API
  version: "1.0.0"
  title: "ZWF API"
basePath: "/api/v1"
schemes:
  - http
  - https
...
paths:
...
  /org/payment_method/:id:
    delete:
      operationId: deleteOrgPaymentMethod
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /org/payment_method/:id/primary:
    post:
      operationId: setOrgPrimaryPaymentMethod
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
...
  /doc_template/{id}:
    get:
      operationId: getDocTemplate
      # security:
      #   - authLoggedInUser: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
    delete:
      operationId: deleteDocTemplate
      # security:
      #   - authAdmin: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /doc_template/{id}/clone:
    put:
      operationId: cloneDocTemplate
      # security:
      #   - authLoggedInUser: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
  /doc_template/{id}/rename:
    post:
      operationId: renameDocTemplate
      # security:
      #   - authClient: []
      parameters:
        - name: id
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...